pyproject.toml
[project]
name = "dicom-mapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "highdicom>=0.27.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pydicom>=3.0.1",
    "simpleitk>=2.5.3",
    "tqdm>=4.67.1",
]

[project.scripts]

dicom-mapper = "dicom_mapper:main"



[build-system]

requires = ["hatchling"]

build-backend = "hatchling.build"
